+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Unreleased
+
+* Added convert to FLAC option for uncompressed files, with undo/redo and optional delete original
+
 # v1.6
 
 * UI Enhancements (FX icons, highlight on waveform seek, new empty states)